FW-369. Move private functions from radio.c into a cc2420-specific library.
No code modifications were introduced, plain copy paste of existing code.
