+++ bsp/chips/cc2420/SConscript
@@ -7,7 +7,7 @@ Import('env')
 
 localEnv = env.Clone()
 
-source = ['radio.c']
+source = ['radio.c', 'cc2420.c']
 cc2420 = localEnv.Object(
    source=source,
 )
